hunted down some bugs. Got one in Extrude (silly one!) and one in ISelector (missing unregistered_from_grid in destructor).
introduced badNormals to AdaptSurfaceGridToCylinder. Thanks to them one can specify the maximal deviation of a faces normal to the cylinders main-axis. If a vertex is connected to a face with a bad normal it is regarded as a rim-vertex.

Fixed tetrahedralization issues by choosing new command-line options for tetgen (-YY was very useful).

Added marker points.