Major rework on ElemDisc, Coupling procedure and UserData:
CoupledElemDiscs are removed from the assembling process, since they are not needed anymore. Now, all ElemDiscs provide the possibility of Data export and import by itself. This gives a simpler interface, since only a single base class is used. 
The Assembling process has been restructured. Now, at the beginning of the assembling, all scheduled ElemDiscs of a subset are grouped and the assembling is performed in a single loop over all elements of the subset. Furthermore, imports and exports are detected by the DataEvaluator and computed at the time, when they are needed.
The IUserData has been modified. It now derives from the more general IPData base class. This class is intended to support three kinds of data: 1. Constant data, 2. Data depending on time and space and 3. Data provided by other ElemDiscs. The Constant data is evaluated before the loop over all elements is performed. The position dependent data clearly on each element. The data coming from other disc is evaluated on each element, and in case of jacobian computation also the derivatives of the data w.r.t. the ElemDisc local unknowns are computed. In case of jacobian computation, these derivatives together with the linearization of the defect w.r.t. the data is used to compute the inter-ElemDisc coupling matrices. This can be used to assemble exact jacobians, even in case of coupled discretizations.