added "delete" to lua-bindings. If an object was created in the lua script, it can now be deleted using delete(obj).
The page-container now works and can be used to store attachments. A compile-option is however still missing. Such an option should allow to choose the container-type for attachment storage.
The page-container avoids reallocation and instead allocates junks of memory up front. Those memory junks are managed in pages.
Lookup of an entry involves integer / and % operations and is thus slower than a lookup in a std::vector. Growing a container however does not involve copying or deleting data.
Please note: If the page container is used, a considerable memory and runtime overhead is introduced for very small grids (~10 elements).

Added a method which logs all active attachments and the sizes of associated containers (total occupied memory).