Proudly presenting: the MessageHub.
The MessageHub is a helper class which provides a flexible way to register callbacks for given message-ids, which then can be called through the MessageHub::post_message method.
A callback can be a normal function pointer or a pointer to a class member function.
Each message-id can easily be associated with a specific message-type. The only constraint is, that this message-type has to derive from MessageHub::IMessage.
The MessageHub can me found at "common/util/message_hub.h"

Included is a small test case in test_bridge (see MessageHubTest class in bridges/test_bridge.cpp)