Docu: renamed documentation files for lib_algebra, lib_disc, lib_grid
Renaming all the documentation files of the separate libraries used in
ug4 to be in line with the new documentation of ug4.

TODO: adjust the Doxyfiles for generation of lib_*'s separate
documentation to pick up the *.doxygen files. (located in ug4-root ->
doxy_config_lib*.txt). Until then they are ignored and this separate
and optional documentation is somehow useless.