DoFManager now supports partial grid redistribution. Therefore added methods begin_parallel_redistribution and end_parallel_redistribution to MGDoFDistribution. When parallel_redistribution is active, the SurfaceDofDistribution and the LevelDoFDistribution behave slightly different on element creation and deletion. Furthermore a post-processing is executed in those classes when parallel redistribution ends (vmasters are removed from interfaces, etc)
Removed freeze method from the dof-manager.

During distribution, only the parts of a grid which are sent to other processes are serialized again. The part which stays on the local process won't be serialized, instead, all elements which won't stay are simply erased. This has been introduced some revisions ago, but was temporarily disabled again since the DoFManger was not yet prepared for this. This new behavior should lead to a severe performance benefit when only small parts of a grid are to be sent to another process, e.g. during redistribution after adaptive refinement.