[docu] replaced custom table of contents on most pages with autogenerated
Since Doxygen 1.8 isn't that new and shiny any more, the
`\tableofcontents` command and Markdown support (`[TOC]`) is acctual
useable. This command generates a table of contents box on HTML pages
automatically from all `\section`, `\subsection` and `\subsubsection`
commands on a `\page`.
It is much more maintainable as manual listing and referencing all
sections of a page.