Reintegrated branch 'parallel_coarsening'. Parallel adaptive multigrid refiner, DistributedGridManager and DoF-Managers have been reworked so that parallel coarsening works rather good now (tested for mpirun -n 2 ugshell -ex adaptive_mg/moving_front.lua -redistributionSteps 2 -numTimeSteps 14).
A problem surfaced which has not yet been fixed: When ghost-elements are removed during coarsening, surface-dof-indices will be distributed on their parents. However, since the ghost elements did not have surface-dof-indices, no restriction can be performed and the associated vector entries stay uninitialized. The testrun specified above has those problems starting at time-step 14.

One more note: The parent-object-type of an element in a multi-grid can now be queried even if the parent element is not located on the querying process.