Added a safety-check to domain_distribution_util.lua (used e.g. by util.CreateAndDistributeDomain): An error is raised if the number of elements in the distribution-level is smaller than the number of target processes.
adaptive_henry_compare.lua now also works in parallel. A simple bisection distribution is used, so that the globally-refined grid parts and the adaptively refined parts match on each process. This is required to evaluate GlobalGridFunctionNumberData. Therefore, no redistribution is performed during the simulation. This is fine, since the global-solution should be at least as expensive as the adaptive solution.