Introducing functionality for a  posteriori error estimation.
It is now possible to add assembling methods for an error estimator of your choice to any elem_disc.
Example implementations are already realized in detail for convection_diffusion_fv1 (the convection part might not be entirely correct though), neumann_boundary_fv1 and inner_boundary_fv1.
In order to calculate element-wise error indicators from the estimator values assembled in the elem_discs, you can use one of the pre-implemented ErrEstData classes in err_est_data.h or a derived class of your own thereof.

This feature is still in the testing phase. Thus all comments and bug reports are welcome (for now, hehe).