added checkpoint_util.lua: scaling MPI-I/O checkpointing methods and LUA tables. example: (from util.SolveLinearTimeProblemRestart)
-- example writing a checkpoint (std is every 10s) 
	util.WriteCheckpointIntervallic(u, time, {time=time, step=step})
	
-- example reading a checkpoint
	local restart = util.HasParamOption("-restart") 
	if restart then
		cp = util.ReadCheckpoint(u)
		
		-- re-get our additional data
		time = cp.myData.time
		step = cp.myData.step
	end
	
Note: The used SaveToFile/ReadFromFile functions for u
assume that u has exactly the same structure as in the checkpoint.
That means approxSpace, grid and so on have to be the same
This will normally exclude adaptive calculations from restarting,
unless you add code which also reads/writes the adapted grid. 
See also util.GetStdCheckpointData for variables which are
automatically checked to be the same between runs.