Fix for creation of nan coordinates by ProjectionHandler::handle_new_vertex()
The new position was calculated as 0*oldPos + 1*avgParentPos on subsets where no refinement projector was set.
Obviously, however, and sadly, oldPos was uninitialized as it is just to be set in this method.
So in the event where it had a +/-nan value, even multiplication by 0 would not make that go away.
This led to very hard to reproduce nan values in the vertex coordinates on some systems
or to very reproducible nan errors on others (I guess due to carpeting the memory with 1 prior to allocation).
This error had persisted for several months (Nov 17, 2016).