Made future time available for time step preparation
Until now, the prepare_timestep() method of the IDomainDiscretization interface was unaware of the time the next step was assembled for, it knew only of the (old) time given in the previous solution. However, some elem discs may (and do in fact) need information about the step size itself. So the future time is passed down to them as well now, starting from the time discretization.