corrected small ugliness in ThetaTimeStep
The methods assemble_linear() and assemble_rhs() in ThetaTimeStep time discretization was using rhs vector as (unknown) next time step solution rather than the old solution.
This should not have caused any errors as the new solution should not be used in any linear assembling. Still, it is not a good idea and might lead to unexpected behavior, e.g., if the rhs vector has values outside the admissible range for solutions.