Added matrix overlap creation with new overlap interfaces.
The function CreateOverlap extends a matrix by adding all matrix
connections to from associated slave nodes to master nodes. Those new
nodes are stored in so called Master-Overlap-Interfaces. Those are
accessible through the matrix's AlgebraLayout.

Nodes connected to slaves which have not had copies on the process of
associated masters are added to a so called Slave-Overlap-Interface.
Again accessible through the matrix's AlgebraLayout.

Please note: The implementation is still work in progress!