Enhanced the Raster class. Kernels can now be executed on entries.
Provided a 'Count', 'Sum', and 'Blur' kernel.
The 'Raster' class now also provides a 'blur' method.

Added a file 'raster_util.lua' which can automatically determine the
dimension of a .asc file. It creates Value or Matrix callbacks for
raster-lookups matching a required dimension, independent of the
dimension of the underlying raster file.