Added -copyScript option to ugsubmit
When activated, this option will cause the script specified after the -ex argument to be copied to the execution directory AND the execution to take place on the copied script (instead of the original one).
This allows to put batch jobs in the queue and then adjust the original scripts without consequences for the queued jobs.
Furthermore, this serves as documentation of the simulation carried out in the execution directory.