Added a lightweight wrapper to ugshell_main and an option ALTERNATE_MAIN
Using ALTERNATE_MAIN, an alternative main file may be specified. This
can be a simple wrapper of the form

    \#include "ug_shell/ugshell_main.h"
    int main (int argc, char** argv)
    {
    	return ugshell_main (argc, argv);
    }

ALTERNATE_MAIN can be specified as a cmake argument or in a CMakeLists.txt
of a plugin. Here is some documentation on how to use ALTERNATE_MAIN:

A lightweight wrapper for ugshell_main is provided in ug_shell/main.cpp.
It simply calls ugshell_main with the provided command-line arguments.

If you'd like to do something before ugshell_main is executed, you may
write your own alternate_main.cpp and compile ug with the
\code
	-DALTERNATE_MAIN=PATH_TO_YOUR_MAIN/alternate_main.cpp
\endcode
option.

Alternatively, you may also create a plugin or enhance your existing one. To this end,
add an 'alternate_main.cpp' file and a CMakeLists.txt to your plugin which contains the line:
\code
	set(ALTERNATE_MAIN "${CMAKE_CURRENT_SOURCE_DIR}/alternate_main.cpp" CACHE PATH "Wrapper to ugshell_main." FORCE)
\endcode

In that case, the alternate_main.cpp is automatically activated as soon as you enable your plugin.

If you want to return to the original main, please deactivate the plugin in question and
once call:
\code
	cmake -DUNSET_ALTERNATE_MAIN=ON .
\endcode