Rework of global grid distribution quality measure
The total quality has previously been computed as the minimum of
the individual level qualities. This is bad in adaptive refinement cases,
where higher levels can contain very few elements.

Now, the total quality measure is computed just like the level-wise
measure, i.e., as the fraction
    (optimal expected compute time) / (actual expected compute time)
and normalized to the interval [0,1].