Fix in SectionContainer
In adaptive simulations with redistribution it could happen that
section end iterators were left to point to elements that no longer existed.
If new elements were created in that location, the section iteration loops
could be affected in such a way that discretizations did not assemble the complete
matrices / defects etc. A very subtle, but serious bug that has persisted since the
beginning of UG4.
The fix now makes sure that the end iterators are always valid.