Added ExpectedErrorMarkingStrategy
This marking strategy marks elements for refinement until the expected error
of the refined solution reaches a specified tolerance.
It refines elements in the order given by a (global) sorting with respect to
the elements' error indicator values.
Parallelization is somewhat crude: THe sorting is performed on the highest-rank
process (as this is usually the one with the most memory left).