Fix for bad implementation of hanging node constraints for linear problems
The implementation no longer distinguishes the linear from the non-linear case.
Both handle the constrained nodes by effectively taking them out of the computation
by making sure that the defect and corrections of constrained DoFs are always zero
and their value in the solution does not change.
To make this possible in the linear assembling case, the initial solution had to be
introduced as an additional argument to the assembling function.

ATTENTION: Unlike before, it is now necessary to call adjust_solution() once after
the solver is finished to enforce the hanging node constraints in linear problems.
Otherwise, the DoFs of hanging nodes will have value 0 in the solution.