added nullptr to SmartPtr class so that a smart pointer can be initialized and compared with a null-pointer.
Since this can lead to ambiguous interpretation of NULL this should be replaced by nullptr.
Note NULL can be evaluated to __null, 0, nullptr and is not typesafe

caution with integration: while for ugcore NULL was replaced by nullptr, plugins were not checked if they contain NULL where SPNULL was meant.